backend/requirements.txt
airtable-python-wrapper==0.15.3
certifi==2022.6.15.1
charset-normalizer==2.1.1
idna==3.3
isort==5.10.1
numpy==1.23.3
Pillow==9.2.0
pycodestyle==2.9.1
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
toml==0.10.2
transliterate==1.10.2
urllib3==1.26.12
Wand==0.6.10
